requirements.txt
discord.py[speed] @ git+https://github.com/Rapptz/discord.py
uvloop
asyncpg
jishaku
psutil
git+https://github.com/PythonistaGuild/mystbin.py
tabulate

